upadate setelah terhubung vm

+++ app/Http/Controllers/BagianpklController.php
@@ -75,16 +75,17 @@ class BagianpklController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
-    {
-        try {
-        $data = Bagianpkl::findOrFail($id);
-        $data->delete();
-        return redirect()->route('bagianpkl.index')->with('success','Data berhasil dihapus');
+{
+    $data = Bagianpkl::findOrFail($id);
 
-        } catch (QueryException $e) {
-            if($e->getCode() == "23000") {
-                return redirect()->route('bagianpkl.index')->with('error','Data tidak dapat dihapus! data ini masih digunakan di data penempatan PKL');
-            }
-        }
+    if ($data->penempatanpkl()->exists()) {
+        return redirect()->route('bagianpkl.index')
+            ->with('error', 'Data tidak dapat dihapus! Data ini masih digunakan di data penempatan PKL');
     }
+
+    $data->delete();
+
+    return redirect()->route('bagianpkl.index')
+        ->with('success', 'Data berhasil dihapus');
+}
 }
